requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.8
async-generator==1.10
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.5
contourpy==1.0.5
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
Django==4.0.6
django-admin-interface==0.19.1
django-autocomplete-light==3.9.4
django-colorfield==0.7.2
django-filter==22.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-jazzmin==2.5.0
docopt==0.6.2
docxcompose==1.3.5
docxtpl==0.16.4
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.1.1
fastjsonschema==2.16.2
fonttools==4.37.4
gunicorn==20.1.0
h11==0.13.0
idna==3.4
ipykernel==6.16.0
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.16.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
kiwisolver==1.4.4
loguru==0.6.0
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.6.1
matplotlib-inline==0.1.6
mistune==2.0.4
mysqlclient==2.1.1
nbclient==0.7.0
nbconvert==7.2.1
nbformat==5.6.1
nest-asyncio==1.5.6
notebook==6.4.12
numpy==1.22.3
openpyxl==3.0.10
outcome==1.2.0
packaging==21.3
pandas==1.5.0
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
pipreqs==0.4.11
prometheus-client==0.14.1
prompt-toolkit==3.0.31
psutil==5.9.2
pure-eval==0.2.2
pycparser==2.21
Pygments==2.13.0
PyMySQL==1.0.2
pyparsing==3.0.9
PyPDF2==2.11.1
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.21.0
pytz==2022.2.1
pywin32==304
pywinpty==2.0.8
pyzmq==24.0.1
qtconsole==5.3.2
QtPy==2.2.1
requests==2.28.1
scikit-learn==1.1.2
scipy==1.9.3
selenium==4.4.3
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.5.1
terminado==0.16.0
threadpoolctl==3.1.0
tinycss2==1.1.1
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
transliterate==1.10.2
trio==0.21.0
trio-websocket==0.9.2
tzdata==2022.1
urllib3==1.26.12
wcwidth==0.2.5
webdriver-manager==3.8.3
webencodings==0.5.1
widgetsnbextension==4.0.3
win32-setctime==1.1.0
wsproto==1.2.0
yarg==0.1.9
